Frontend/requirements.txt
Flask==2.1.1
python-dotenv
requests
